# API_SPEC.json
{
  "swagger": "2.0",
  "info": {
    "title": "AI交易终端API接口规范",
    "description": "股票交易分析系统六大核心模块API接口设计规范",
    "version": "1.0.0",
    "contact": {
      "name": "技术团队",
      "email": "tech@ai-trading-terminal.com"
    }
  },
  "host": "api.ai-trading-terminal.com",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer {token}"
    }
  },
  "security": [{
    "BearerAuth": []
  }],
...
  "paths": {
    "/chip/distribution": {
      "get": {
        "tags": ["chip"],
        "summary": "获取单只股票筹码分布数据",
        "description": "返回指定股票指定时间段内的筹码分布、集中度、主力成本等核心数据",
        "parameters": [
          {
            "name": "stockCode",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "股票代码（如 SH600000、SZ000001）"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "开始日期（yyyy-MM-dd），默认近 30 天"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "description": "结束日期（yyyy-MM-dd），默认当天"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/ChipDistributionResponse"
            }
          },
          "400": {
            "description": "参数无效",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "BaseResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 200
        },
        "msg": {
          "type": "string",
          "example": "success"
        },
        "requestId": {
          "type": "string",
          "example": "req-1234567890abcdef"
        },
        "timestamp": {
          "type": "integer",
          "example": 1735689600000
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 400
        },
        "msg": {
          "type": "string",
          "example": "参数无效/缺失"
        },
        "requestId": {
          "type": "string",
          "example": "req-1234567890abcdef"
        },
        "timestamp": {
          "type": "integer",
          "example": 1735689600000
        }
      }
    },
...
    "ChipDistributionResponse": {
      "allOf": [
        {
          "$ref": "#/definitions/BaseResponse"
        },
        {
          "type": "object",
          "properties": {
            "data": {
              "type": "object",
              "properties": {
                "stockCode": {
                  "type": "string",
                  "example": "SH600000"
                },
                "stockName": {
                  "type": "string",
                  "example": "浦发银行"
                },
                "date": {
                  "type": "string",
                  "example": "2026-01-04"
                },
                "chipDistribution": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "price": {
                        "type": "integer",
                        "example": 850
                      },
                      "chipRatio": {
                        "type": "number",
                        "example": 0.05
                      },
                      "holderCount": {
                        "type": "integer",
                        "example": 12000
                      }
                    }
                  }
                },
                "chipConcentration": {
                  "type": "number",
                  "example": 0.89
                },
                "mainCostPrice": {
                  "type": "integer",
                  "example": 820
                },
                "supportPrice": {
                  "type": "integer",
                  "example": 790
                },
                "resistancePrice": {
                  "type": "integer",
                  "example": 900
                },
                "chipPeakInfo": {
                  "type": "object",
                  "properties": {
                    "peakPrice": {
                      "type": "integer",
                      "example": 830
                    },
                    "peakRatio": {
                      "type": "number",
                      "example": 0.65
                    },
                    "isSinglePeak": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      ]
    },
...
  },
...
}